requirements.txt
django==1.9
psycopg2==2.7.1
djangorestframework==3.6.3
django-filter==1.0.4
Pillow==4.1.1
openpyxl==2.4.8